# src/frontend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "VenueAxe.Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5280/"
    }
  ],
  "paths": {
...
    "/api/admin/bookings/venue/{venueId}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/waivers/templates/venue/{venueId}": {
      "get": {
        "tags": [
          "Waivers"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverTemplateDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverTemplateDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverTemplateDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/booking-config/venue/{venueId}": {
      "get": {
        "tags": [
          "BookingConfig"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfigDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfigDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfigDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "BookingConfig"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingConfigRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingConfigRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfigDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfigDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfigDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/venue/{venueId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSummaryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSummaryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{id}/role": {
      "put": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/users/{id}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/venues": {
      "get": {
        "tags": [
          "Venues"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VenueDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VenueDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VenueDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Venues"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVenueRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVenueRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVenueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VenueDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/venues/{id}": {
      "get": {
        "tags": [
          "Venues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VenueDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Venues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVenueRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVenueRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVenueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VenueDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/lanes/venue/{venueId}": {
      "get": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaneDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaneDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LaneDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Lanes"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLaneRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLaneRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLaneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaneDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookingConfigDto": {
        "required": [
          "id",
          "venueId",
          "minPartySize",
          "maxPartySize",
          "slotDurationsMinutes",
          "turnaroundBufferMinutes",
          "pricingModel",
          "basePriceCents",
          "peakPriceCents",
          "depositType",
          "depositAmountCents",
          "editorThemeJson",
          "customFieldsJson",
          "packagesJson",
          "discountRulesJson",
          "bookingTypesJson",
          "addonsJson",
          "cancellationPolicy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "minPartySize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "maxPartySize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "slotDurationsMinutes": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          "turnaroundBufferMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pricingModel": {
            "$ref": "#/components/schemas/PricingModel"
          },
          "basePriceCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "peakPriceCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "depositType": {
            "$ref": "#/components/schemas/DepositType"
          },
          "depositAmountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "editorThemeJson": {
            "type": "string"
          },
          "customFieldsJson": {
            "type": "string"
          },
          "packagesJson": {
            "type": "string"
          },
          "discountRulesJson": {
            "type": "string"
          },
          "bookingTypesJson": {
            "type": "string"
          },
          "addonsJson": {
            "type": "string"
          },
          "cancellationPolicy": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "BookingDto": {
        "required": [
          "id",
          "venueId",
          "bookingReference",
          "status",
          "guestFirstName",
          "guestLastName",
          "guestEmail",
          "guestPhone",
          "partySize",
          "startTime",
          "endTime",
          "totalAmountCents",
          "paidAmountCents",
          "paymentStatus",
          "assignedLaneNumbers",
          "signedWaiverCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "bookingReference": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "guestFirstName": {
            "type": "string"
          },
          "guestLastName": {
            "type": "string"
          },
          "guestEmail": {
            "type": "string"
          },
          "guestPhone": {
            "type": "string"
          },
          "partySize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "paidAmountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "paymentStatus": {
            "type": "string"
          },
          "assignedLaneNumbers": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          "signedWaiverCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "bookingTypeId": {
            "type": [
              "null",
              "string"
            ]
          },
          "discountAmountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32",
            "default": 0
          },
          "appliedDiscountCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "squarePaymentId": {
            "type": [
              "null",
              "string"
            ]
          },
          "balanceDueCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "BookingStatus": {
        "type": "integer"
      },
...
      "CreateUserRequest": {
        "required": [
          "venueId",
          "email",
          "fullName",
          "role",
          "password"
        ],
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "CreateVenueRequest": {
        "required": [
          "name",
          "slug",
          "addressLine1",
          "city",
          "state",
          "postalCode",
          "phone",
          "email"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": [
              "null",
              "string"
            ]
          },
          "addressLine1": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "timezone": {
            "type": "string",
            "default": "America/New_York"
          },
          "currency": {
            "type": "string",
            "default": "USD"
          }
        }
      },
      "DepositType": {
        "type": "integer"
      },
...
      "PricingModel": {
        "type": "integer"
      },
...
      "UpdateBookingConfigRequest": {
        "required": [
          "minPartySize",
          "maxPartySize",
          "slotDurationsMinutes",
          "turnaroundBufferMinutes",
          "pricingModel",
          "basePriceCents",
          "peakPriceCents",
          "depositType",
          "depositAmountCents",
          "editorThemeJson",
          "customFieldsJson",
          "packagesJson",
          "discountRulesJson",
          "bookingTypesJson",
          "addonsJson",
          "cancellationPolicy"
        ],
        "type": "object",
        "properties": {
          "minPartySize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "maxPartySize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "slotDurationsMinutes": {
            "type": "array",
            "items": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          "turnaroundBufferMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pricingModel": {
            "$ref": "#/components/schemas/PricingModel"
          },
          "basePriceCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "peakPriceCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "depositType": {
            "$ref": "#/components/schemas/DepositType"
          },
          "depositAmountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "editorThemeJson": {
            "type": "string"
          },
          "customFieldsJson": {
            "type": "string"
          },
          "packagesJson": {
            "type": "string"
          },
          "discountRulesJson": {
            "type": "string"
          },
          "bookingTypesJson": {
            "type": "string"
          },
          "addonsJson": {
            "type": "string"
          },
          "cancellationPolicy": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateUserRoleRequest": {
        "required": [
          "role"
        ],
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        }
      },
...
      "UserRole": {
        "type": "integer"
      },
      "UserSummaryDto": {
        "required": [
          "id",
          "venueId",
          "email",
          "fullName",
          "role",
          "isActive",
          "lastLoginAt",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "venueId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "isActive": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "WaiverTemplateDto": {
        "required": [
          "id",
          "venueId",
          "versionNumber",
          "title",
          "bodyTextMarkdown",
          "sha256Hash"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "versionNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "bodyTextMarkdown": {
            "type": "string"
          },
          "sha256Hash": {
            "type": "string"
          }
        }
      }
    }
  },
...
}